refactor: intense refactoring
